API: move decorators & AllTracker w/ helper functions to separate files
