+++ src/pytools/api/__init__.py
@@ -2,5 +2,7 @@
 Basic tools for API development, supporting documentation, deprecation,
 and run-time validation.
 """
+from ._alltracker import *
 from ._api import *
+from ._decorators import *
 from ._doc_validator import *
